cambios en la estructura, alineaciones del framework, detalles visuales, manejo de endpoint de pago, cambios alrededor del carrito de compras
