+++ app/payment-method.tsx
@@ -1,116 +1,147 @@
 import React, { useState } from "react";
-import { View, Text, StyleSheet, TouchableOpacity, Image, Alert } from "react-native";
+import { View, Text, StyleSheet, TouchableOpacity, Image, Alert, ScrollView } from "react-native";
 import { Ionicons } from "@expo/vector-icons";
 import { router, useLocalSearchParams } from "expo-router"; // Importamos el hook para acceder a los parámetros
 import { createOrder } from "@/service/OrderService";
+import { createBill } from "@/service/BillService"; // Importamos el nuevo endpoint
 import { makeOrderPayload } from "@/lib/OrderUtils";
 import { useCart } from "@/contexts/CartContext";
+import AsyncStorage from '@react-native-async-storage/async-storage';
 
 export default function Payment() {
-  const { cartItems, updateQuantity, removeFromCart } = useCart();
+  const { cartItems, updateQuantity, removeFromCart, setCartItems } = useCart();
   const { billName = "", nit = "" } = useLocalSearchParams(); // Aseguramos que billName y nit tengan un valor por defecto
   const [selectedMethod, setSelectedMethod] = useState("QR"); // Estado para el método de pago seleccionado
 
   return (
-    <View style={styles.container}>
-      {/* Botones de selección de método de pago */}
-      <View style={styles.paymentMethods}>
-        <TouchableOpacity
-          style={[
-            styles.methodButton,
-            selectedMethod === "QR" && styles.activeMethod,
-          ]}
-          onPress={() => setSelectedMethod("QR")}
-        >
-          <Text style={styles.methodText}>Código QR</Text>
-        </TouchableOpacity>
+    <ScrollView contentContainerStyle={styles.scrollContainer}>
+      <View style={styles.container}>
+        {/* Botones de selección de método de pago */}
+        <View style={styles.paymentMethods}>
+          <TouchableOpacity
+            style={[
+              styles.methodButton,
+              selectedMethod === "QR" && styles.activeMethod,
+            ]}
+            onPress={() => setSelectedMethod("QR")}
+          >
+            <Text style={styles.methodText}>Código QR</Text>
+          </TouchableOpacity>
 
-        <TouchableOpacity
-          style={[
-            styles.methodButton,
-            selectedMethod === "Card" && styles.activeMethod,
-          ]}
-          onPress={() => setSelectedMethod("Card")}
-        >
-          <Text style={styles.methodText}>Tarjeta</Text>
-        </TouchableOpacity>
-      </View>
-
-      {/* Vista del código QR si está seleccionado */}
-      {selectedMethod === "QR" && (
-        <Image
-          source={require("../assets/images/qr_code.png")} // Cambia la ruta según tu carpeta
-          style={styles.image}
-        />
-      )}
-      {/* Vista de detalles del usuario */}
-      <View style={[styles.card, styles.userDetails]}>
-        <View style={styles.detailRow}>
-          <Ionicons name="information-circle-outline" size={24} color="black" />
-          <Text style={styles.detailText}>
-            {billName ? billName : "Nombre no proporcionado"}{" "}
-            {/* Valor predeterminado */}
-          </Text>
+          <TouchableOpacity
+            style={[
+              styles.methodButton,
+              selectedMethod === "Card" && styles.activeMethod,
+            ]}
+            onPress={() => setSelectedMethod("Card")}
+          >
+            <Text style={styles.methodText}>Tarjeta</Text>
+          </TouchableOpacity>
         </View>
 
-        <View style={styles.detailRow}>
-          <Ionicons name="pricetag-outline" size={24} color="black" />
-          <Text style={styles.detailText}>
-            {nit ? nit : "NIT/CI no proporcionado"} {/* Valor predeterminado */}
-          </Text>
+        {/* Vista del código QR si está seleccionado */}
+        {selectedMethod === "QR" && (
+          <Image
+            source={require("../assets/images/qr_code.png")} // Cambia la ruta según tu carpeta
+            style={styles.image}
+          />
+        )}
+        {/* Vista de detalles del usuario */}
+        <View style={[styles.card, styles.userDetails]}>
+          <View style={styles.detailRow}>
+            <Ionicons name="information-circle-outline" size={24} color="black" />
+            <Text style={styles.detailText}>
+              {billName ? billName : "Nombre no proporcionado"}{" "}
+              {/* Valor predeterminado */}
+            </Text>
+          </View>
+
+          <View style={styles.detailRow}>
+            <Ionicons name="pricetag-outline" size={24} color="black" />
+            <Text style={styles.detailText}>
+              {nit ? nit : "NIT/CI no proporcionado"} {/* Valor predeterminado */}
+            </Text>
+          </View>
         </View>
-      </View>
 
-      {/* Total */}
-      <View style={styles.totalContainer}>
-        <Text style={styles.totalText}>Total:</Text>
-        <Text style={styles.totalAmount}>Bs. 104</Text>
-      </View>
+        {/* Total */}
+        <View style={styles.totalContainer}>
+          <Text style={styles.totalText}>Total:</Text>
+          <Text style={styles.totalAmount}>Bs. 104</Text>
+        </View>
 
-      {/* Botón de "Pedir" */}
-      <TouchableOpacity style={styles.orderButton}
-        onPress={() => {
-          console.log("pressed pay button");  
-          Alert.alert(
-            "Confirmar pedido",
-            "¿Estás seguro de que deseas realizar el pedido?",
-            [
-              {
-                text: "Cancelar",
-                onPress: () => console.log("Prompt cancelled"),
-                style: "cancel"
-              },
-              {
-                text: "OK",
-                onPress: async () => {
-                  console.log("cartitems", cartItems);
-                  // Create the order
-                  const orderPayload = makeOrderPayload(1, cartItems);
-                  console.log("makeorderpayload", orderPayload);
-                  try {
+        {/* Botón de "Pedir" */}
+        <TouchableOpacity style={styles.orderButton}
+          onPress={() => {
+            console.log("pressed pay button");  
+            Alert.alert(
+              "Confirmar pedido",
+              "¿Estás seguro de que deseas realizar el pedido?",
+              [
+                {
+                  text: "Cancelar",
+                  onPress: () => console.log("Prompt cancelled"),
+                  style: "cancel"
+                },
+                {
+                  text: "OK",
+                  onPress: async () => {
+                    console.log("cartitems", cartItems);
                     // Create the order
-                    const response = await createOrder(orderPayload);
-                    Alert.alert("Pedido realizado", "Tu pedido ha sido registrado exitosamente.");
-                    // Redirect user to the home page
-                    // TODO: Clear the cart after the order is created
-                    router.push('/');
-                  } catch (error) {
-                    console.error("Error creating order:", error);
-                    Alert.alert("Error", "Hubo un error al registrar tu orden. intenta nuevamente.");
-                  } 
+                    const orderPayload = makeOrderPayload(1, cartItems);
+                    console.log("makeorderpayload", orderPayload);
+                    try {
+                      // Create the order
+                      const orderResponse = await createOrder(orderPayload);
+                      console.log("Order response:", orderResponse);
+
+                      if (!orderResponse.payload.orderId) {
+                        throw new Error("Order ID is missing in the response");
+                      }
+
+                      // Ensure billName and nit are strings
+                      const billNameString = Array.isArray(billName) ? billName.join(", ") : billName;
+                      const nitString = Array.isArray(nit) ? nit.join(", ") : nit;
+
+                      // Create the bill
+                      const billPayload = {
+                        orderId: orderResponse.payload.orderId,
+                        billName: billNameString || "Nombre no proporcionado",
+                        nit: nitString || "NIT/CI no proporcionado",
+                        totalCost: 104, // Reemplaza esto con el cálculo real del costo total
+                      };
+                      const billResponse = await createBill(billPayload);
+                      console.log("Bill response:", billResponse);
+
+                      Alert.alert("Pedido realizado", "Tu pedido ha sido registrado exitosamente.");
+
+                      // Clear the cart
+                      setCartItems([]);
+                      await AsyncStorage.removeItem('cartItems');
+
+                      // Redirect user to the home page
+                      router.push('/');
+                    } catch (error) {
+                      console.error("Error creating order or bill:", error);
+                      Alert.alert("Error", "Hubo un error al registrar tu orden. intenta nuevamente.");
+                    }
+                  }
                 }
-              }
-            ]
-          );
-        }}
-        >
-        <Text style={styles.orderButtonText}>Pedir</Text>
-      </TouchableOpacity>
-    </View>
+              ]
+            );
+          }}
+          >
+          <Text style={styles.orderButtonText}>Pedir</Text>
+        </TouchableOpacity>
+      </View>
+    </ScrollView>
   );
 }
 
 const styles = StyleSheet.create({
+  scrollContainer: {
+    flexGrow: 1,
+  },
   container: {
     flex: 1,
     padding: 20,
