+++ contexts/CartContext.tsx
@@ -10,70 +10,14 @@ interface CartContextProps {
   updateQuantity: (id: number, delta: number) => void;
   removeFromCart: (id: number) => void;
   setMenuItems: (items: Product[]) => void;
+  setCartItems: (items: Product[]) => void; // Añadido
 }
 
 const CartContext = createContext<CartContextProps | undefined>(undefined);
 
 export const CartProvider = ({ children }: { children: ReactNode }) => {
   const [cartItems, setCartItems] = useState<Product[]>([]);
-  const [menuItems, setMenuItems] = useState<Product[]>([
-    {
-        id: 1,
-        name: 'Onigiris de Atún',
-        description: 'Deliciosos triángulos de arroz rellenos de atún fresco, sazonados con un toque de salsa de soya y envueltos en una capa de alga nori crujiente.',
-        price: 25,
-        image: 'https://images.pond5.com/pixel-sushi-vector-illustration-isolated-illustration-155825087_iconm.jpeg',
-        quantity: 10
-      },
-      {
-        id: 2,
-        name: 'Cheesecake de Uvas',
-        description: 'Un postre delicioso y fresco, perfecto para cualquier ocasión.',
-        price: 30,
-        image: 'https://assets.tmecosys.com/image/upload/t_web767x639/img/recipe/vimdb/230649.jpg',
-        quantity: 5
-      },
-      {
-        id: 3,
-        name: 'Tacos de Pollo',
-        description: 'Tacos de pollo con guacamole y salsa de chipotle.',
-        price: 40,
-        image: 'https://www.vvsupremo.com/wp-content/uploads/2017/06/Chicken-Tacos-900x570-sRGB.jpg',
-        quantity: 20
-      },
-      {
-        id: 4,
-        name: 'Pizza de Pepperoni',
-        description: 'Pizza de pepperoni con queso mozzarella y salsa de tomate.',
-        price: 50,
-        image: 'https://encrypted-tbn0.gstatic.com/images?q=tbn:ANd9GcSIoXjS-sXqWGIsMTB_m3av-Oh-Fgi93hBrzg&s',
-        quantity: 8
-      },
-      {
-        id: 5,
-        name: 'Hamburguesa Clásica',
-        description: 'Hamburguesa con carne de res, lechuga, tomate, cebolla y queso cheddar.',
-        price: 35,
-        image: 'https://img.freepik.com/fotos-premium/foto-stock-hamburguesa-clasica-aislada-blanco_940723-217.jpg',
-        quantity: 15
-      },
-      {
-        id: 6,
-        name: 'Té Helado',
-        description: 'Té helado de limón, perfecto para refrescarte en un día caluroso.',
-        price: 15,
-        image: 'https://imag.bonviveur.com/te-helado.jpg',
-        quantity: 25
-      },
-      {
-        id: 7,
-        name: 'Pastel de Chocolate',
-        description: 'Un pastel de chocolate esponjoso y delicioso, perfecto para los amantes del chocolate.',
-        price: 30,
-        image: 'https://i.pinimg.com/736x/42/36/b1/4236b10d070cb898106d84a6f2fa4a2c.jpg',
-        quantity: 12
-      }
-    ]);
+  const [menuItems, setMenuItems] = useState<Product[]>([]);
 
   useEffect(() => {
     const loadCart = async () => {
@@ -177,7 +121,7 @@ export const CartProvider = ({ children }: { children: ReactNode }) => {
   };
 
   return (
-    <CartContext.Provider value={{ cartItems, menuItems, addToCart, updateQuantity, removeFromCart, setMenuItems }}>
+    <CartContext.Provider value={{ cartItems, menuItems, addToCart, updateQuantity, removeFromCart, setMenuItems, setCartItems }}>
       {children}
     </CartContext.Provider>
   );
