+++ models/Bill.ts
@@ -1,0 +1,6 @@
+export interface Bill {
+    orderId: number;
+    billName: string;
+    nit: string;
+    totalCost: number;
+}
